handle exception in feedback

+++ backend/feedback/views.py
@@ -2,12 +2,17 @@ from django.http import HttpResponse
 from django.db import connection
 import json
 
-
 # Feedback Table is this
 # id | date_time | student_id | title | feedback_description
 
 # http://127.0.0.1:8000/feedback/?hostel=rajiv
+from rest_framework import status
+
+
 def get_feedback(request):
+    if request.method != 'GET':
+        return HttpResponse(content='only get request allowed', status=status.HTTP_400_BAD_REQUEST)
+
     hostel = request.GET.get('hostel', None)
     if hostel is None:
         with connection.cursor() as cursor:
@@ -34,13 +39,29 @@ def get_feedback(request):
 
 
 def add_feedback(request):
-    student_id = request.POST.get('student_id', '1111')
-    title = request.POST.get('title', 'review')
-    feedback_description = request.POST.get('feedback_description', 'not good')
+    if request.method != 'POST':
+        return HttpResponse(content='only post request allowed', status=status.HTTP_400_BAD_REQUEST)
+
+    student_id = request.POST.get('student_id', None)
+    title = request.POST.get('title', None)
+    feedback_description = request.POST.get('feedback_description', None)
     # raw_query = "INSERT INTO public.feedback (id , date_time, student_id, title, feedback_description) " \
     #             "VALUES (DEFAULT, NOW()::TIMESTAMP(0), '%s', '%s', '%s')" % (student_id, title, feedback_description)
     #
     # print(raw_query)
+    if (student_id is None) or (title is None) or (feedback_description is None):
+        return HttpResponse(content="all data not provided : student_id, title, feedback_description",
+                            status=status.HTTP_400_BAD_REQUEST)
+
+    with connection.cursor() as cursor:
+        cursor.execute("""
+        SELECT EXISTS( SELECT * FROM public.student WHERE enrollment_no = %s);
+        """, (student_id,))
+        row = cursor.fetchone()
+    if not row[0]:
+        return HttpResponse(content="student with this enrollment_no does not exists",
+                            status=status.HTTP_400_BAD_REQUEST)
+
     with connection.cursor() as cursor:
         cursor.execute("""
         INSERT INTO public.feedback (id , date_time, student_id, title, feedback_description) 
@@ -52,7 +73,18 @@ def add_feedback(request):
 
 
 def del_feedback(request):
-    _id = int(request.POST.get('id', 1))
+    if request.method != 'POST':
+        return HttpResponse(content='only post request allowed', status=status.HTTP_400_BAD_REQUEST)
+
+    _id = request.POST.get('id', None)
+    if _id is None:
+        return HttpResponse(content="all data not provided : id", status=status.HTTP_400_BAD_REQUEST)
+
+    try:
+        _id = int(_id)
+    except ValueError:
+        return HttpResponse(content='id can not convert into int', status=status.HTTP_400_BAD_REQUEST)
+
     with connection.cursor() as cursor:
         cursor.execute("""
         DELETE FROM public.feedback WHERE id = %s ;
